- generate demo transaction - get users transactions

+++ database/migrations/2021_11_18_064523_create_transactions_table.php
@@ -20,7 +20,7 @@ class CreateTransactionsTable extends Migration
             $table->string('plan_name');
             $table->string('driver');
             $table->string('reference');
-            $table->integer('amount');
+            $table->decimal('amount');
             $table->text('currency');
             $table->timestamps();
         });
